+++ src/backend/Node.c
@@ -3,10 +3,10 @@
 
 typedef struct node {
 	gpointer name;
-	int id;
+	gsize id;
 } *Node;
 
-Node make_node(gpointer name, int id){
+Node make_node(gpointer name, gsize id){
 	Node node = g_malloc(sizeof(struct node));
 
 	node->name = name;
@@ -20,11 +20,11 @@ gpointer get_name_node(Node node){
 	return node->name;
 }
 
-int get_id_node(Node node){
+gsize get_id_node(Node node){
 	return node->id;
 }
 
-void set_id_node(Node node, int id){
+void set_id_node(Node node, gsize id){
 	node->id = id;
 }
